Write data descriptor signature to be like Info-ZIP
Previously pigz did not write a data descriptor in order to follow the
PKWare appnote, which does not call for a signature there.  However,
Info-ZIP zip has been putting a signature there since 1996, and there
may be no other software that writes data descriptors, which are only
required when streaming zip file output.  Futhermore, the most recent
PKWare appnote says that there may be a signature, and that conforming
decoders need to allow for its presence or absence.  So pigz now does
include a signature, with the assurance that zip decoders will be able
to handle it, and in fact are likely to expect it.