Add --rsyncable functionality.
This adds the --rsyncable option, analogous to the gzip patch that
provides the same option.  The implementation here is different, as
noted in the comments.  --rsyncable inserts resynchronization points
(many of them) in the compressed data, where the location of the
points is a function of recent input data.  Then when a large file
that was compressed has small portions modified and is then
recompressed (both times with the --rsyncable option), then only
portions of the compressed data will need to be retransmitted when
using rsync, since most of it will remain unchanged.