Remove volatile in yarn.c.
A user indicated that volatile was needed for the threads global in
yarn.c for an HP system.  If that is the case, then there is a bug
in the HP pthread library.  Testing by other users suggested that
there is not a bug.  In any case, pigz will not try to work around
such bugs with volatile, since volatile in general won't solve such
problems anyway.