Have rsyncable compression be independent of threading.
This patch allows pigz with no threading or with threading make the
same compression requests, both with and without the --independent
option.  While necessary, this patch is not sufficient.  A patch to
zlib is also needed, which will be in zlib version 1.2.5.3 and
later.  Even with all of these patches, the output may still differ
for compression levels one through three.