Do not require that -d precede associated -N, -n, -T options.
The default use of the header name and time is different for
compression and decompression, being put in both for compression
and use neither for decompression.  Before, to get the proper
interpretation of -N with -d, it had to follow -d.  -dN would
decompress using the name and time, but -Nd would be the same as
-d.  This commit makes the behavior independent of the position of
the -d.