Do not abort on inflate data error -- continue to process files.
This introduces try/catch/throw for error handling throughout pigz.
Each thread has its own try stack for error handling.  If a throw
makes it to the top try in a thread, then the entire program is
aborted.  Data errors caught during decoding and decompression
permit the process to continue with subsequent files.