Add -M, --time option to store or restore the modification time.
This replaces the previous -M for --maxsplits with -J. (Hopefully
no one was using -M. Probably not.) This fills a hole left by the
change of -n, --no-name to also not store/restore the modification
time. Now -nM will store/restore the modification time, but not
the name. The full set of combinations is -n for neither, -N for
both, -nM for just the time, and -Nm for just the name. Note that
the order is important. -n or -N will overwrite the effect of a
preceding -m or -M. The default for compression is -N, and the
default for decompression is -n. So -N or -n can be left off for
those cases, respectively.