Add run-time zlib version check to handle weak-linking case.
This allows for weak linking with a zlib library whose version is
older than the zlib header file used to compile pigz. The version
is checked at run time, and deflatePending() is not called if the
linked library does not have it. This commit adds a check as well
when invoked for the earliest acceptable version of zlib (1.2.3).