Process options first.
Change command line processing to go through all of the options
and associated parameters first, and then go through the command
line again, processing the file names using the resolved options.
In addition, if "-" follows a "--", then consider the "-" to be
a file name instead of stdin. Both of these changes are to be more
compatible with gzip, and are not compatible with the previous
versions of pigz, which would process the options and file names
in the sequence they appear.