Improve compatibility for cross-platform testing.
Avoid warning for ignoring chown() return value on some glibc's,
and change compiler from cc to gcc. Sometimes the (void) before
chown() is not sufficient to signal the programmer's intent to
ignore the return value. (Thank you return-value police.) This
commit kills that gcc warning just for copymeta(). Using pragma
then required ignoring a warning for versions of gcc that don't
recognize it (when using -Wall). Finally, gcc is used instead of
cc for platforms that don't make that equivalency.