Don't delete a .zip file with more than one entry.
To be consistent with gzip, pigz will delete the compressed file
once it has completed decompression. This applies to both gzip and
zip files. However it only decompresses the first entry in a zip
file. If there is more than one entry in the zip file, pigz would
delete it even though not all of the zip contents were extracted.
This commit prevents the deletion of the zip file if entries after
the first are detected.