Do not set mtime when reading from pipe
Currently `tar --gzip/-z/-Igzip` produces deterministic tarballs from
the same input, because it sets a 0 modification time in the gzip
header.

On the other hand, `tar -Ipigz` does not produce deterministic output,
because it uses the current time as modification time.

It would be good to be consistent with gzip here.